+++ EssentialApp/EssentialAppTests/Helpers/FeedLoaderStub.swift
@@ -1,0 +1,20 @@
+//
+//  FeedLoaderStub.swift
+//  EssentialAppTests
+//
+//  Created by Prabhat Tiwari on 15/05/22.
+//
+
+import EssentialFeed
+
+class FeedLoaderStub: FeedLoader {
+   private let result: FeedLoader.Result
+   
+   init(result: FeedLoader.Result) {
+       self.result = result
+   }
+   
+   func load(completion: @escaping (FeedLoader.Result) -> Void) {
+       completion(result)
+   }
+}
